requirements.txt
fastapi
httpx
python-dotenv
uvicorn
psycopg2-binary
